Some changes

+++ EMS/View/StartForm.cs
@@ -38,10 +38,20 @@ namespace EMS.View
             detailLayoutPanel.RowStyles[0].Height = 90;
             detailLayoutPanel.RowStyles[1].Height = 0;
 
-            topicBindingSource.DataSource = new List<Topic>(1);
+            topicBindingSource.DataSource = new List<Topic>(0);
             topicDataListView.DataSource = topicBindingSource;
             disableTopicButtons();
+            topicLabel.Text = "";
+
+            activityDataListView.DataSource = activityBindingSource;
+            disableActivityButtons();
+
+            eventSearchTextBox.Hide();
+            topicSearchTextBox.Hide();
+            activitySearchTextBox.Hide();
         }
+
+
 
         private void addEventButton_Click(object sender, EventArgs e)
         {
@@ -72,21 +82,64 @@ namespace EMS.View
 
         private void deleteEventButton_Click(object sender, EventArgs e)
         {
-            Event ev = (Event)eventListView.SelectedObject;
-            if (eM.deleteEvent(ev))
+            DialogResult result = MessageBox.Show("Are you sure you want to delete event?", "Event Managment System", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
             {
-                eventBindingSource.Remove(ev);
+                Event ev = (Event)eventListView.SelectedObject;
+                if (eM.deleteEvent(ev))
+                {
+                    eventBindingSource.Remove(ev);
+                }
             }
+        }
+
+        private void addToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            addEventButton.PerformClick();
+        }
+
+        private void editToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            if (eventListView.SelectedIndex != -1)
+            {
+                editEventButton.PerformClick();
+            }
+            else
+            {
+                MessageBox.Show("You have to select event.", "Event Managment System", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+        }
+
+        private void deleteToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            if (eventListView.SelectedIndex != -1)
+            {
+                deleteEventButton.PerformClick();
+            }
+            else
+            {
+                MessageBox.Show("You have to select event.", "Event Managment System", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+        }
+
+        private void eventButtonEnabled(bool enabled)
+        {
+            editEventButton.Enabled = deleteEventButton.Enabled = enabled;
+            addTopicButton.Enabled = enabled;
         }
 
         private void eventListView_SelectedIndexChanged(object sender, EventArgs e)
         {
+            
+
             if (eventListView.SelectedIndex == -1)
             {
                 eventButtonEnabled(false);
                 disableTopicButtons();
                 detailLayoutPanel.RowStyles[0].Height = 90;
                 detailLayoutPanel.RowStyles[1].Height = 0;
+                topicBindingSource.Clear();
+                activityBindingSource.Clear();
             }
             else
             {
@@ -103,23 +156,50 @@ namespace EMS.View
                 topicBindingSource.DataSource = ev.Topics.Where(t => t.Active == true).ToList();
             }
 
+            topicDataListView.SelectedIndex = -1;
+            activityDataListView.SelectedIndex = -1;
+
         }
 
-        private void eventButtonEnabled(bool enabled)
+        private void eventListView_SizeChanged(object sender, EventArgs e)
         {
-            editEventButton.Enabled = deleteEventButton.Enabled = enabled;
+            DataListView listView = sender as DataListView;
+            if (listView != null)
+            {
+                float totalColumnWidth = 0;
+
+                // Get the sum of all column tags
+                for (int i = 0; i < listView.Columns.Count; i++)
+                    totalColumnWidth += Convert.ToInt32(listView.Columns[i].Tag);
+
+                // Calculate the percentage of space each column should 
+                // occupy in reference to the other columns and then set the 
+                // width of the column to that percentage of the visible space.
+                for (int i = 0; i < listView.Columns.Count; i++)
+                {
+                    float colPercentage = (Convert.ToInt32(listView.Columns[i].Tag) / totalColumnWidth);
+                    listView.Columns[i].Width = (int)(colPercentage * listView.ClientRectangle.Width);
+                }
+            }
         }
 
-        private void disableTopicButtons()
+        private void eventSearchTextBox_TextChanged(object sender, EventArgs e)
         {
-            editTopicButton.Enabled = addTopicButton.Enabled = deleteTopicButton.Enabled = false;
+            searchEvents();
         }
 
-        private void topicButtonEnabled(bool enabled)
+        private void searchEvents()
         {
-            addTopicButton.Enabled = true;
-            editTopicButton.Enabled = deleteTopicButton.Enabled = viewTopicButton.Enabled = enabled;
+            eventListView.UseFiltering = true;
+            eventListView.ModelFilter = new ModelFilter(x =>
+            {
+                var ev = x as Event;
+                return x != null && (ev.Title.ToLower().Contains(eventSearchTextBox.Text.ToLower()));
+            });
         }
+
+
+
 
         private void addTopicButton_Click(object sender, EventArgs e)
         {
@@ -151,34 +231,15 @@ namespace EMS.View
 
         private void deleteTopicButton_Click(object sender, EventArgs e)
         {
-            Topic t = (Topic)topicDataListView.SelectedObject;
-            if (tM.deleteTopic(t))
+            DialogResult result = MessageBox.Show("Are you sure you want to delete topic?", "Event Managment System", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
             {
-                topicBindingSource.Remove(t);
+                Topic t = (Topic)topicDataListView.SelectedObject;
+                if (tM.deleteTopic(t))
+                {
+                    topicBindingSource.Remove(t);
+                }
             }
-        }
-
-        private void topicDataListView_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            if (topicDataListView.SelectedIndex == -1)
-            {
-                topicButtonEnabled(false);
-            }
-            else
-            {
-                topicButtonEnabled(true);
-            }
-
-        }
-
-        private void eventSearchTextBox_TextChanged(object sender, EventArgs e)
-        {
-            searchEvents();
-        }
-
-        private void topicSearchTextBox_TextChanged(object sender, EventArgs e)
-        {
-            searchTopics();
         }
 
         private void viewTopicButton_Click(object sender, EventArgs e)
@@ -187,55 +248,6 @@ namespace EMS.View
             new TopicDetailViewForm(this, (Topic)topicDataListView.SelectedObject, aM).ShowDialog();
             this.Show();
 
-        }
-
-        private void searchEvents()
-        {
-            eventListView.UseFiltering = true;
-            eventListView.ModelFilter = new ModelFilter(x =>
-            {
-                var ev = x as Event;
-                return x != null && (ev.Title.ToLower().Contains(eventSearchTextBox.Text.ToLower()));
-            });
-        }
-
-        private void searchTopics()
-        {
-            topicDataListView.UseFiltering = true;
-            topicDataListView.ModelFilter = new ModelFilter(x =>
-            {
-                var t = x as Topic;
-                return x != null && (t.Title.ToLower().Contains(topicSearchTextBox.Text.ToLower()));
-            });
-        }
-
-        private void addToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            addEventButton.PerformClick();
-        }
-
-        private void editToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            if (eventListView.SelectedIndex != -1)
-            {
-                editEventButton.PerformClick();
-            }
-            else
-            {
-                MessageBox.Show("You have to select event.", "Event Managment System", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            }
-        }
-
-        private void deleteToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            if (eventListView.SelectedIndex != -1)
-            {
-                deleteEventButton.PerformClick();
-            }
-            else
-            {
-                MessageBox.Show("You have to select event.", "Event Managment System", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            }
         }
 
         private void addToolStripMenuItem1_Click(object sender, EventArgs e)
@@ -274,6 +286,17 @@ namespace EMS.View
             }
         }
 
+        private void disableTopicButtons()
+        {
+            addTopicButton.Enabled = editTopicButton.Enabled = deleteTopicButton.Enabled = viewTopicButton.Enabled = false;
+        }
+
+        private void topicButtonEnabled(bool enabled)
+        {
+            editTopicButton.Enabled = deleteTopicButton.Enabled = viewTopicButton.Enabled = enabled;
+            addActivityButton.Enabled = enabled;
+        }
+
         private void viewToolStripMenuItem_Click(object sender, EventArgs e)
         {
             if (topicDataListView.SelectedIndex != -1)
@@ -286,9 +309,105 @@ namespace EMS.View
             }
         }
 
+        private void topicDataListView_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            if (topicDataListView.SelectedIndex == -1)
+            {
+                topicButtonEnabled(false);
+                activityButtonEnabled(false);
+                activityBindingSource.Clear();
+                topicLabel.Text = "";
+            }
+            else
+            {
+                topicButtonEnabled(true);
+                activityButtonEnabled(false);
+                Topic t = (Topic)topicDataListView.SelectedObject;
+                activityBindingSource.DataSource = t.Activities.Where(a => a.Active == true).ToList();
+                topicLabel.Text = "\"" + t.Title + "\":";
+            }
+
+            activityDataListView.SelectedIndex = -1;
+
+        }
+
         private void topicDataListView_SizeChanged(object sender, EventArgs e)
         {
             // Don't allow overlapping of SizeChanged calls
+            if (!resizing)
+            {
+                // Set the resizing flag
+                resizing = true;
+
+                DataListView listView = sender as DataListView;
+                if (listView != null)
+                {
+                    float totalColumnWidth = 0;
+
+                    // Get the sum of all column tags
+                    for (int i = 0; i < listView.Columns.Count; i++)
+                        totalColumnWidth += Convert.ToInt32(listView.Columns[i].Tag);
+
+                    // Calculate the percentage of space each column should 
+                    // occupy in reference to the other columns and then set the 
+                    // width of the column to that percentage of the visible space.
+                    for (int i = 0; i < listView.Columns.Count; i++)
+                    {
+                        float colPercentage = (Convert.ToInt32(listView.Columns[i].Tag) / totalColumnWidth);
+                        listView.Columns[i].Width = (int)(colPercentage * listView.ClientRectangle.Width);
+                    }
+                }
+            }
+
+            // Clear the resizing flag
+            resizing = false;
+        }
+
+        private void topicSearchTextBox_TextChanged(object sender, EventArgs e)
+        {
+            searchTopics();
+        }        
+
+        private void searchTopics()
+        {
+            topicDataListView.UseFiltering = true;
+            topicDataListView.ModelFilter = new ModelFilter(x =>
+            {
+                var t = x as Topic;
+                return x != null && (t.Title.ToLower().Contains(topicSearchTextBox.Text.ToLower()));
+            });
+        }
+
+
+
+
+
+
+
+        private void disableActivityButtons()
+        {
+            addActivityButton.Enabled = editActivityButton.Enabled = deleteActivityButton.Enabled = viewActivityButton.Enabled = false;
+        }
+        
+        private void activityButtonEnabled(bool enabled)
+        {
+            editActivityButton.Enabled = deleteActivityButton.Enabled = viewActivityButton.Enabled = enabled;
+        }
+
+        private void activityDataListView_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            if(activityDataListView.SelectedIndex == -1)
+            {
+                activityButtonEnabled(false);
+            }
+            else
+            {
+                activityButtonEnabled(true);
+            }
+        }
+
+        private void activityDataListView_SizeChanged(object sender, EventArgs e)
+        {
             if (!resizing)
             {
                 // Set the resizing flag
